backend/requirements.txt
Bottleneck
brotlipy==0.7.0
certifi==2022.12.7
cffi
chardet==5.0.0
charset-normalizer
click
cryptography
decorator
et-xmlfile==1.1.0
Flask
Flask-Cors
idna
importlib-metadata
itsdangerous
Jinja2
lml==0.1.0
MarkupSafe
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
numexpr
numpy
openpyxl==3.0.10
packaging
pandas==1.4.3
pycparser
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
PyJWT
pymongo==3.12.0
pyOpenSSL
pyparsing
PySocks
python-dateutil
python-dotenv
pytz
requests
six
texttable==1.6.4
Unidecode
urllib3
validators
Werkzeug
wget==3.2
xlrd==2.0.1
xlwt==1.3.0
zipp